+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add new lowest rate functions for Shipments, Orders and Pickups
 - Add new lowest smartrate functions for Shipments
 - Adds OS details to the User-Agent header
+- Add support for webhook secrets
 
 ## v2.3.0 (2022-04-13)